+++ src/Aks/Aks/Commands/NewAzureRmAks.cs
@@ -496,12 +496,12 @@ namespace Microsoft.Azure.Commands.Aks
                     defaultAgentPoolProfile.NodeLabels.Add(key.ToString(), NodePoolLabel[key].ToString());
                 }
             }
-            if (this.IsParameterBound(c => c.NodePoolTags))
+            if (this.IsParameterBound(c => c.NodePoolTag))
             {
                 defaultAgentPoolProfile.Tags = new Dictionary<string, string>();
-                foreach (var key in NodePoolTags.Keys)
+                foreach (var key in NodePoolTag.Keys)
                 {
-                    defaultAgentPoolProfile.Tags.Add(key.ToString(), NodePoolTags[key].ToString());
+                    defaultAgentPoolProfile.Tags.Add(key.ToString(), NodePoolTag[key].ToString());
                 }
             }
             if (this.IsParameterBound(c => c.AvailabilityZone))