+++ src/Aks/Aks/Commands/NewOrUpdateAgentPoolBase.cs
@@ -43,9 +43,9 @@ namespace Microsoft.Azure.Commands.Aks
         public string Mode { get; set; }
 
         [Parameter(Mandatory = false, HelpMessage = "Node pool labels used for building Kubernetes network.")]
-        public Hashtable NodeLabels { get; set; }
+        public Hashtable NodeLabel { get; set; }
 
         [Parameter(Mandatory = false, HelpMessage = "The tags to be persisted on the agent pool virtual machine scale set.")]
-        public Hashtable Tags { get; set; }
+        public Hashtable Tag { get; set; }
     }
 }