+++ src/Aks/Aks/ChangeLog.md
@@ -22,8 +22,8 @@
 * Added parameter `-NodeOsSKU` for `New-AzAksCluster` and parameter `-OsSKU` for `New-AzAksNodePool`
 * Added parameter `-Mode` for `New-AzAksNodePool` and `Update-AzAksNodePool`
 * Added property `-NodeImageVersion` for the output of `Get-AzAksNodePool`[#19893]
-* Added parameter `-NodePoolLabel` for `Set-AzAksCluster`, `-NodeLabels` for `New-AzAksNodePool` and `Update-AzAksNodePool`
-* Added parameter `-NodePoolTags` for `New-AzAksCluster` and `Set-AzAksCluster`, `-Tags` for `New-AzAksNodePool` and `Update-AzAksNodePool`
+* Added parameter `-NodePoolLabel` for `Set-AzAksCluster`, `-NodeLabel` for `New-AzAksNodePool` and `Update-AzAksNodePool`
+* Added parameter `-NodePoolTag` for `New-AzAksCluster` and `Set-AzAksCluster`, `-Tag` for `New-AzAksNodePool` and `Update-AzAksNodePool`
 
 ## Version 5.0.1
 * Upgraded AutoMapper to Microsoft.Azure.PowerShell.AutoMapper 6.2.2 with fix [#18721]